stock_checker/requirements.txt
streamlit
pandas
plotly
yfinance
numpy
google-genai
python-dotenv
pytest
